Move Z download handling to separate manager, and try download from Z first with fallback to torrent
